refactor: rename test classes and update package structure for clarity
